+++ src/commands/pickFuncProcess.ts
@@ -8,8 +8,8 @@ import ps = require('ps-node');
 import { isNullOrUndefined } from 'util';
 import * as vscode from 'vscode';
 import { IActionContext, UserCancelledError } from 'vscode-azureextensionui';
-import { validateFuncCoreToolsInstalled } from '../commands/createNewProject/validateFuncCoreToolsInstalled';
 import { extensionPrefix, isWindows } from '../constants';
+import { validateFuncCoreToolsInstalled } from '../funcCoreTools/validateFuncCoreToolsInstalled';
 import { localize } from '../localize';
 import { getFuncExtensionSetting } from '../ProjectSettings';
 import { cpUtils } from '../utils/cpUtils';
