+++ test/createNewProject.test.ts
@@ -11,9 +11,9 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { DialogResponses, IActionContext, TestUserInput } from 'vscode-azureextensionui';
 import { createNewProject } from '../src/commands/createNewProject/createNewProject';
-import { funcToolsInstalled } from '../src/commands/createNewProject/validateFuncCoreToolsInstalled';
 import { ProjectLanguage } from '../src/constants';
 import { ext } from '../src/extensionVariables';
+import { funcToolsInstalled } from '../src/funcCoreTools/validateFuncCoreToolsInstalled';
 import { dotnetUtils } from '../src/utils/dotnetUtils';
 import * as fsUtil from '../src/utils/fs';
 import { validateVSCodeProjectFiles } from './initProjectForVSCode.test';
